fix: decimal prices, i18n for trip cards and detail page
- Change priceEur from Int to Float in Prisma schema + Zod validation
- Fix PriceItemsEditor: use rawPrices string state so decimals (19,95) can be typed
- Replace hardcoded "Pro Person", "Sichere dir..." with t() calls
- Add bookingTagline translation key (DE/EN)
- formatDate now uses useLocale() instead of hardcoded de-DE
- Rename "Slug" label to "Slug (URL-Pfad)" in admin trip form
- Fix list bullet alignment using native list-disc + marker: instead of mt hacks

+++ src/lib/validations.ts
@@ -61,7 +61,7 @@ export const tripSchema = z.object({
   bookingStatus: z.enum(bookingStatuses).optional().default("AVAILABLE"),
   departureDate: z.string().min(1),
   returnDate: z.string().min(1),
-  priceEur: z.number().int().positive(),
+  priceEur: z.number().positive(),
   destination: z.string().min(1),
   imageUrl: z.string().optional().default(""),
   translations: z.object({
